merging neccton in bioptimod merge

+++ testcase/create_bc_nc.py
@@ -25,12 +25,12 @@ def create_bc_nc(test):
     tmask   =  M.variables['tmask'].data[0,:,:,:].astype(bool).copy()
     M.close()
 
-    D3=np.ones((1,jpk,jpj,jpi),np.float)
-    D2=np.ones((1,jpj,jpi),np.float)
+    D3=np.ones((1,jpk,jpj,jpi),np.float64)
+    D2=np.ones((1,jpj,jpi),np.float64)
 
 # Creating bounmask.nc
 
-    index=np.zeros((jpk,jpj,jpi),np.int)
+    index=np.zeros((jpk,jpj,jpi),int)
 
     waterpoints = 0;
 
@@ -41,7 +41,7 @@ def create_bc_nc(test):
                     waterpoints += 1
                     index[jk,jj,ji] = waterpoints
 
-    index_inv=np.zeros((waterpoints,3),np.int)
+    index_inv=np.zeros((waterpoints,3),int)
 
     waterpoints = 0;
 
@@ -99,7 +99,7 @@ def create_bc_nc(test):
                 if tmask[0,jj,ji]:
                     atm_idxt += 1
 
-    atm_index=np.ones((atm_idxt),np.float)
+    atm_index=np.ones((atm_idxt),np.float64)
 
     atm_idxt = 0;
 
@@ -133,7 +133,7 @@ def create_bc_nc(test):
 
     filein.close()
 
-    D2=np.ones((jpj,jpi),np.float)
+    D2=np.ones((jpj,jpi),np.float64)
 
 
     for date in CO2_DATE:
@@ -170,7 +170,7 @@ def create_bc_nc(test):
             if (jk ==0 ):
                 tin_idxt += 1
 
-    riv_index=np.ones((tin_idxt),np.int)
+    riv_index=np.ones((tin_idxt),int)
 
     tin_idxt = 0;
 
@@ -183,12 +183,12 @@ def create_bc_nc(test):
                 riv_index[tin_idxt]=index[jk,jj,ji]
                 tin_idxt += 1
 
-    riv_N1p=np.zeros(tin_idxt,dtype=float); riv_N1p[0]=0.35*10**(-5)
-    riv_N3n=np.zeros(tin_idxt,dtype=float); riv_N3n[0]=0.2*10**(-3)
-    riv_N5s=np.zeros(tin_idxt,dtype=float); riv_N5s[0]=1.0*10**(-4)
-    riv_O3c=np.zeros(tin_idxt,dtype=float); riv_O3c[0]=0.35
-    riv_O3h=np.zeros(tin_idxt,dtype=float); riv_O3h[0]=0.01
-    riv_O2o=np.zeros(tin_idxt,dtype=float); riv_O2o[0]=0.002
+    riv_N1p=np.zeros(tin_idxt,dtype=np.float64); riv_N1p[0]=0.35*10**(-5)
+    riv_N3n=np.zeros(tin_idxt,dtype=np.float64); riv_N3n[0]=0.2*10**(-3)
+    riv_N5s=np.zeros(tin_idxt,dtype=np.float64); riv_N5s[0]=1.0*10**(-4)
+    riv_O3c=np.zeros(tin_idxt,dtype=np.float64); riv_O3c[0]=0.35
+    riv_O3h=np.zeros(tin_idxt,dtype=np.float64); riv_O3h[0]=0.01
+    riv_O2o=np.zeros(tin_idxt,dtype=np.float64); riv_O2o[0]=0.002
 
     for date in TIN_DATE:
         # Create RIV file
@@ -226,7 +226,7 @@ def create_bc_nc(test):
         if ( (ji==1) | (jj==1) ) | ( (ji==jpi-2) | (jj==jpj-2) ):
             gib_idxt += 1
 
-    gib_index=np.ones((gib_idxt),np.int)
+    gib_index=np.ones((gib_idxt),int)
 
     gib_idxt = 0;
